Rename events module to matrix
